requirements.txt
ipykernel
pandas
numpy
openpyxl
langchain
langchain-community
langchain-chroma
pydantic
langchain-openai
python-dotenv
tiktoken
chromadb
langgraph
langchain-text-splitters
fastapi
uvicorn
networkx
lxml

